cli: rename 'cluster example' command to 'create cluster'
